r1902: Make drbdmeta compile without warnings on amd64

+++ user/drbdtool_common.c
@@ -26,7 +26,7 @@ char* ppsize(char* buf, size_t size)
 		size = size >> 10;
 		base++;
 	}
-	sprintf(buf,"%d %cB",size,units[base]);
+	sprintf(buf,"%lu %cB",(unsigned long)size,units[base]);
 
 	return buf;
 }
